requirements.txt
tensorflow==2.1.*
ffmpy==0.2.*
librosa==0.7.*
numpy==1.18.*
Pillow==7.0.*
opencv-python==4.1.2.*
PyYAML==5.*
pandas==1.0.*
click==7.*
matplotlib==3.2.*
scikit-learn==0.22.*
jupyter==1.0.*
natsort==7.0.*
click-option-group==0.3.*